refactor: Add callback to dismiss overlay.
The app contains an overlay to indicate progress in fetching data when the user's username is changed or entered for the first time. Add a callback interface to allow UserSetFragment and UserStudiedFragment to communicate to MainActivity that the fetch attempt is complete. Implement the interface method synchronized in MainActivity because both fragments call it.