feat: Add DB column in Sets for term_count and Listener to CustomViewPager.
The ActionBar in the DetailActivity is updated to show the user the number of the card they're on and the total number of cards in the set. Needed to add term_count to the Set database table and pass it to DetailActivity in the Intent. DetailActivity uses this data to populate the ActionBar in OnCreate and also, via a listener added to the CustomViewPager. Note: getItem() method on ViewPagerAdapter wasn't the right solution - it is called twice at first and has trouble with the last position as well.